feat: add @wagmi/core dependency and implement user redirection hooks in various components
