+++ package.json
@@ -25,6 +25,7 @@
     "@radix-ui/react-switch": "^1.1.3",
     "@radix-ui/react-tooltip": "^1.1.8",
     "@tanstack/react-query": "^5.67.2",
+    "@wagmi/core": "^2.16.7",
     "@yudiel/react-qr-scanner": "^2.2.1",
     "alchemy-sdk": "^3.5.4",
     "axios": "^1.8.2",
