+++ src/app/register/page.tsx
@@ -2,8 +2,11 @@
 
 import Navbar from "@/components/layout/register/navbar";
 import { RegistrationForm } from "@/components/layout/register/registration-form";
+import { useRedirectUserToWallet } from "@/hooks/useRedirectUserToWallet";
 
 export default function RegisterPage() {
+  useRedirectUserToWallet();
+
   return (
     <div>
       <Navbar />
